requirements.txt
openai
fastapi
uvicorn
sqlite3
python-dotenv
